+++ examples/dowser.rs
@@ -20,7 +20,8 @@ fn main() {
 	let now = Instant::now();
 	let files: Vec<PathBuf> = Dowser::default()
 		.with_path("/usr/share")
-		.into_vec_filtered(|p| Some(EXT) == Extension::try_from2(p));
+		.filter(|p| Some(EXT) == Extension::try_from2(p))
+		.collect();
 
 	println!("Search took {} seconds.", now.elapsed().as_millis() as f64 / 1000.000);